Rename file based on PR feedback
